db_setup/requirements.txt
aperturedb>=0.4.0
python-dotenv>=1.0.0
Pillow>=10.0.0
faker>=19.0.0
email-validator>=2.0.0
requests>=2.31.0
